mbedtls: cast sockfd to int when storing in net_context fd
The mbedtls server and client bio setup assigned lws_sockfd_type
directly to mbedtls_net_context.fd, which is an int. On Win64
lws_sockfd_type is SOCKET (a 64-bit UINT_PTR), so MSVC warns C4244
"possible loss of data", which fails the build under /WX.

Add an explicit (int) cast, matching the gnutls and openhitls backends
which already cast the same sockfd the same way.

+++ lib/tls/mbedtls/mbedtls-client.c
@@ -113,7 +113,7 @@ lws_ssl_client_bio_create(struct lws *wsi)
 		lws_mbedtls_set_alpn(conn->ctx, alpn_comma);
 	}
 
-	conn->net.MBEDTLS_PRIVATE_V30_ONLY(fd) = wsi->desc.sockfd;
+	conn->net.MBEDTLS_PRIVATE_V30_ONLY(fd) = (int)wsi->desc.sockfd;
 	mbedtls_ssl_set_bio(&conn->ssl, &conn->net, lws_plat_mbedtls_net_send, lws_plat_mbedtls_net_recv, NULL);
 
 	return 0;
